+++ astropy/io/ascii/connect.py
@@ -99,7 +99,7 @@ def write_latex(table, filename, **kwargs):
 io_registry.register_writer('latex', Table, write_latex)
 
 
-def is_latex(origin, args, kwargs):
+def is_latex(origin, *args, **kwargs):
     return isinstance(args[0], basestring) and args[0].endswith('.tex')
 
 io_registry.register_identifier('latex', Table, is_latex)
@@ -126,7 +126,7 @@ def write_rdb(table, filename, **kwargs):
 io_registry.register_writer('rdb', Table, write_rdb)
 
 
-def is_rdb(origin, args, kwargs):
+def is_rdb(origin, *args, **kwargs):
     return isinstance(args[0], basestring) and args[0].endswith('.rdb')
 
 io_registry.register_identifier('rdb', Table, is_rdb)